# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "DONE. API",
    "version" : "1.0.0",
    "description" : "Native Apple API. Email/password JWT and native Sign in with Apple. All timestamps UTC; history is server-owned. Durable activity events; APNs transport is not part of this phase."
  },
  "servers" : [ {
    "description" : "Local development; configure PUBLIC_ORIGIN for deployed links.",
    "url" : "http://127.0.0.1:8080"
  } ],
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      },
      "writerCapability" : {
        "description" : "43-character base64url private writer key; never a JWT and never included in the public link.",
        "type" : "http",
        "bearerFormat" : "Private write capability",
        "scheme" : "bearer"
      }
    },
    "schemas" : {
...
      "ApplianceNotificationSettings" : {
        "type" : "object",
        "properties" : {
          "notificationsEnabled" : {
            "type" : "boolean"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer",
            "minimum" : 0
          }
        },
        "additionalProperties" : false,
        "required" : [ "notificationsEnabled", "version" ]
      },
      "ApplianceView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "householdId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "name" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "kind" : {
            "anyOf" : [ {
              "enum" : [ "washer", "dryer", "dishwasher", "oven", "hob", "custom" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "status" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "notificationsEnabled" : {
            "anyOf" : [ {
              "type" : "boolean"
            }, {
              "type" : "null"
            } ]
          },
          "programs" : {
            "anyOf" : [ {
              "items" : {
                "$ref" : "#/components/schemas/ProgramView"
              },
              "type" : "array"
            }, {
              "type" : "null"
            } ]
          },
          "activeSession" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/SessionView"
            }, {
              "type" : "null"
            } ]
          },
          "createdAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "additionalProperties" : false,
        "required" : [ "id", "householdId", "name", "kind", "status", "notificationsEnabled", "programs", "activeSession", "createdAt", "version" ]
      },
      "CatalogItem" : {
        "type" : "object",
        "properties" : {
          "kind" : {
            "anyOf" : [ {
              "enum" : [ "washer", "dryer", "dishwasher", "oven", "hob", "custom" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "title" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "defaultName" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "supportsStopwatch" : {
            "type" : "boolean"
          },
          "needsCollection" : {
            "type" : "boolean"
          },
          "programs" : {
            "anyOf" : [ {
              "items" : {
                "$ref" : "#/components/schemas/Suggestion"
              },
              "type" : "array"
            }, {
              "type" : "null"
            } ]
          },
          "titleKey" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "required" : [ "kind", "title", "defaultName", "supportsStopwatch", "needsCollection", "programs", "titleKey" ]
      },
...
      "CreateAppliance" : {
        "type" : "object",
        "properties" : {
          "kind" : {
            "enum" : [ "washer", "dryer", "dishwasher", "oven", "hob", "custom" ],
            "type" : "string"
          },
          "name" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "householdId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "notificationsEnabled" : {
            "anyOf" : [ {
              "type" : "boolean"
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "required" : [ "kind" ]
      },
...
      "MoveAppliance" : {
        "type" : "object",
        "properties" : {
          "householdId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer",
            "minimum" : 0
          }
        },
        "additionalProperties" : false,
        "required" : [ "householdId", "version" ]
      },
      "MutationResult" : {
        "type" : "object",
        "properties" : {
          "revision" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "additionalProperties" : false,
        "required" : [ "revision" ]
      },
...
      "Problem" : {
        "required" : [ "status", "code", "detail" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer"
          },
          "detail" : {
            "type" : "string"
          }
        }
      },
      "ProgramInput" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer",
            "minimum" : 0,
            "maximum" : 1440
          }
        },
        "additionalProperties" : false,
        "required" : [ "minutes" ]
      },
      "ProgramSnapshot" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "localizationKey" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "isCustom" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "description" : "Immutable value snapshot. name is literal user content or a built-in label; localizationKey is optional. Measured values are separate from the original snapshot.",
        "required" : [ "name", "minutes", "isCustom" ]
      },
      "ProgramView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "name" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "suggested" : {
            "type" : "boolean"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          },
          "applianceId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "localizationKey" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "isCustom" : {
            "type" : "boolean"
          },
          "sortOrder" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "additionalProperties" : false,
        "required" : [ "id", "name", "minutes", "suggested", "version", "applianceId", "localizationKey", "isCustom", "sortOrder" ]
      },
...
      "RenameAppliance" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer",
            "minimum" : 0
          }
        },
        "additionalProperties" : false,
        "required" : [ "name", "version" ]
      },
...
      "SessionView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "applianceId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programName" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "mode" : {
            "anyOf" : [ {
              "enum" : [ "countdown", "stopwatch" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "status" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "startedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "expectedEnd" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "completedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "collectedAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "canceledAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "elapsedSeconds" : {
            "format" : "int64",
            "type" : "integer"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          },
          "serverTime" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "sourceProgramId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "programSnapshot" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramSnapshot"
            }, {
              "type" : "null"
            } ]
          },
          "measuredProgram" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramSnapshot"
            }, {
              "type" : "null"
            } ]
          },
          "timingMode" : {
            "anyOf" : [ {
              "enum" : [ "countdown", "stopwatch" ],
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "isOpen" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "required" : [ "id", "applianceId", "programId", "programName", "minutes", "mode", "status", "startedAt", "expectedEnd", "completedAt", "collectedAt", "canceledAt", "elapsedSeconds", "version", "serverTime", "sourceProgramId", "programSnapshot", "measuredProgram", "timingMode", "isOpen" ]
      },
...
      "Suggestion" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "localizationKey" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "required" : [ "name", "minutes", "localizationKey" ]
      },
...
      "UpdateProgram" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer",
            "minimum" : 1,
            "maximum" : 1440
          },
          "version" : {
            "format" : "int64",
            "type" : "integer",
            "minimum" : 0
          }
        },
        "additionalProperties" : false,
        "required" : [ "name", "minutes", "version" ]
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/appliances" : {
      "get" : {
        "operationId" : "ApplianceController_list_1395660022",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "householdId",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : false,
          "in" : "query"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ApplianceView"
                  },
                  "type" : "array"
                }
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "ApplianceController_create_1395660022",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAppliance"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "201" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApplianceView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appliances/{id}" : {
      "delete" : {
        "operationId" : "ApplianceController_delete_4065617376",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MutationResult"
                }
              }
            }
          }
        }
      },
      "get" : {
        "operationId" : "ApplianceController_get_4065617376",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApplianceView"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "operationId" : "ApplianceController_rename_4065617376",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RenameAppliance"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApplianceView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appliances/{id}/household" : {
      "patch" : {
        "operationId" : "ApplianceController_move_3004726896",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MoveAppliance"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApplianceView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appliances/{id}/notifications" : {
      "patch" : {
        "operationId" : "ApplianceController_notifications_1137391353",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ApplianceNotificationSettings"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApplianceView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appliances/{id}/programs" : {
      "post" : {
        "operationId" : "ApplianceController_program_586500542",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProgramInput"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "201" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProgramView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appliances/{id}/programs/{programId}" : {
      "delete" : {
        "operationId" : "ApplianceController_deleteProgram_3003063688",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "programId",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MutationResult"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "operationId" : "ApplianceController_updateProgram_3003063688",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProgram"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "programId",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProgramView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/catalog" : {
      "get" : {
        "operationId" : "ApplianceController_catalog_434660745",
        "tags" : [ "Appliance" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CatalogItem"
                  },
                  "type" : "array"
                }
              }
            }
          }
        }
      }
    },
...
  }
}